requirements.txt
playwright>=1.40.0
google-generativeai>=0.8.0
jinja2>=3.1.0
readability-lxml>=0.8.1
lxml>=5.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
